requirements.txt
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.6.3
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2018.8.24
chardet==3.0.4
cssselect==1.0.3
cssutils==1.0.2
-e git://github.com/smacker/django-filebrowser-no-grappelli.git@60acccf93561019d5f9d65519b4a9ac3ed0f5298#egg=django_filebrowser_no_grappelli
django-filebrowser-no-grappelli-nph==3.4.9
django-maja-newsletter==0.6.1a11
html2text==2018.1.9
idna==2.7
kombu==3.0.37
lxml==4.2.4
premailer==3.2.0
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
vobject==0.9.6.1
xlrd==1.1.0
xlwt==1.3.0
django-taggit
django-taggit-templatetags